+++ draft-js-emoji-plugin/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## To Be Released
+### Added
+- add emojione as a dependency.
+- change emojiList from emojione lib.
+- config will accept a priorityList, it wil show on top of rest of emojis after user typed ':'.
 
 ## 1.2.3 - 2016-07-06